Finished partial cloud capture, changed architecture

+++ labelling-script.py
@@ -3,19 +3,29 @@ import os
 from mathutils import Vector, Quaternion
 import numpy as np
 import bmesh
+from math import radians
+import pickle
 
 class Plant_Processor:
 
     def __init__(self):
         # USER INPUT: Set active flags for different operations:
-        self.number_of_objects = 1 #total number of plant objects to process
+        self.number_of_objects = 500 #total number of plant objects to process
         self.terrain = True
         self.capture_partial_cloud = True
         self.save_seperated_objects = True
         self.file_path = os.path.join(os.path.expanduser('~'), 'Desktop', 'project', 'deepLeaveSegmentation', 'synth_data', 'original_plant_obj')
         self.save_path = os.path.join(os.path.expanduser('~'), 'Desktop', 'project', 'deepLeaveSegmentation', 'synth_data')
-        self.resolutionX = 1920
-        self.resolutionY = 1080
+        self.resolutionX = 480
+        self.resolutionY = 480
+        self.camera_shots = [(0,0),(20,45),(40,-45)] # A list of camera shots to take of all plants, given as view angle and rotation in degrees
+        self.output_path = os.path.join(os.path.expanduser('~'), 'Desktop', 'project', 'deepLeaveSegmentation', 'synth_data', 'partial_cloud_pickles')
+
+        # More setup operations
+        bpy.context.scene.render.resolution_x = self.resolutionX
+        bpy.context.scene.render.resolution_y = self.resolutionY
+        self.outclouds = None
+        self.outlabels = None
 
     def init_cleanup(self):
         '''Initial clean up of the default scene'''
@@ -23,6 +33,8 @@ class Plant_Processor:
         if bpy.data.objects.get("Cube") is not None:
             bpy.data.objects['Cube'].select_set(True) # Select the default cube object
             bpy.ops.object.delete() # remove the cube
+        bpy.ops.object.empty_add(type='PLAIN_AXES', align='WORLD', location=(0, 0, 0))
+        bpy.data.objects['Camera'].parent = bpy.data.objects['Empty']
 
     def open_and_split(self):
         '''open each generated full plant mesh one by one and split into organs, perform other operations within the loop'''
@@ -47,24 +59,55 @@ class Plant_Processor:
                 self.add_terrain(i)
 
             if self.capture_partial_cloud == True:
-                self.setupCamera() # Here the view can be changed
-                bpy.context.view_layer.update() # Update the view layer once all objects are in place
-                pc_out, labels_out = self.get_partial_pc(i)
+                for shot in self.camera_shots:
+                    self.setupCamera(shot[0], shot[1]) # Here the view is adjusted
+                    bpy.context.view_layer.update() # Update the view layer once all objects are in place
+                    pc_out, labels_out = self.get_partial_pc(i)
+                    self.format_clouds(pc_out, labels_out)
+                    #self.show_point_cloud(pc_out)
+                    pc_out = None
+                    labels_out = None
 
             if self.save_seperated_objects == True:
                 self.save_meshes(i) # saves the split meshes and also deletes them up from the scene
 
-            #self.show_point_cloud(pc_out)
+            self.cleanup() # clear the scene
 
-            self.cleanup()
+            if i%50 == 0:
+                # every now and then export and free up memory
+                self.export_clouds(i/50)
 
     def rename_components(self):
         for obj in list(bpy.data.objects):
             if 'tomato' in obj.name:
                 obj.name = obj.active_material.name
+                obj.name = obj.name.split('.')[0]
 
-    def export_clouds(self):
-        pass
+    def export_clouds(self, id):
+        # Save the collected point clouds in batches
+        cloud_string = '%s_partial_clouds.pickle' %id
+        label_string = '%s_labels.pickle' %id
+        cloud_file = os.path.join(self.output_path, cloud_string)
+        label_file = os.path.join(self.output_path, label_string)
+        with open(cloud_file, 'wb') as f:
+            pickle.dump(self.outclouds,f)
+        with open(label_file, 'wb') as f:
+            pickle.dump(self.outlabels,f)
+
+        # Free up memory and get the next batch
+        self.outclouds = None
+        self.outlabels = None
+
+    def format_clouds(self, cloud, labels):
+        cloud = cloud.reshape(1,cloud.shape[0]*cloud.shape[1],3)
+        labels = labels.reshape(1,labels.shape[0]*labels.shape[1],1)
+
+        if self.outclouds is None and self.outlabels is None:
+            self.outclouds = cloud
+            self.outlabels = labels
+        else:
+            self.outclouds = np.concatenate((self.outclouds,cloud),axis = 0)
+            self.outlabels = np.concatenate((self.outlabels,labels),axis = 0)
 
     def add_terrain(self, i):
         bpy.ops.mesh.landscape_add(refresh=True)
@@ -77,22 +120,25 @@ class Plant_Processor:
         bpy.context.object.ant_landscape.height = 0.05
         bpy.context.object.ant_landscape.lacunarity = 2
         bpy.context.object.ant_landscape.noise_size = 0.5
+        bpy.context.object.name = "Landscape"
+        bpy.ops.mesh.ant_landscape_regenerate()
         bpy.ops.mesh.ant_landscape_refresh()
 
-    def setupCamera(self):
-        # Maybe add some noise to the location and orientation of the camera view
+    def setupCamera(self, view_angle, rotation):
+        # Maybe add some noise to the location and orientation of the camera view? Right now view locations are constant
         bpy.data.objects['Camera'].location[0] = 0
-        bpy.data.objects['Camera'].location[1] = 0
-        bpy.data.objects['Camera'].location[2] = 2
+        bpy.data.objects['Camera'].location[1] = 0 + (view_angle / 500)
+        bpy.data.objects['Camera'].location[2] = 1
         bpy.data.objects['Camera'].rotation_euler[0] = 0
         bpy.data.objects['Camera'].rotation_euler[1] = 0
         bpy.data.objects['Camera'].rotation_euler[2] = 0
-
+        bpy.data.objects['Empty'].rotation_euler[0] = radians(view_angle)
+        bpy.data.objects['Empty'].rotation_euler[2] = radians(rotation)
 
     def cleanup(self):
         all_objects = list(bpy.data.objects)
         for obj in all_objects:
-            if not 'Camera' in obj.name and not 'Light' in obj.name:
+            if not 'Camera' in obj.name and not 'Light' in obj.name and not 'Empty' in obj.name:
                 obj.select_set(True)
                 bpy.ops.object.delete()
 
@@ -131,13 +177,13 @@ class Plant_Processor:
 
         ''' Capture a partial point cloud (simulated depth camera) using ray casting:
         Adapted from https://blender.stackexchange.com/questions/115285/how-to-do-a-ray-cast-from-camera-originposition-to-object-in-scene-in-such-a-w'''
-
         # camera object which defines ray source
         cam = bpy.data.objects['Camera']
+        container = bpy.data.objects['Empty']
         # save current view mode
-        mode = bpy.context.area.type
+        # mode = bpy.context.area.type
         # set view mode to 3D to have all needed variables available
-        bpy.context.area.type = "VIEW_3D"
+        #bpy.context.area.type = "VIEW_3D"
         # get vectors which define view frustum of camera
         frame = cam.data.view_frame(scene=bpy.context.scene)
         topRight = frame[0]
@@ -145,18 +191,19 @@ class Plant_Processor:
         bottomLeft = frame[2]
         topLeft = frame[3]
 
+        label_dict = {'Landscape':0, 'leaves':1, 'stems_1':2, 'trunk':3}
+
         # setup vectors to match pixels
         xRange = np.linspace(topLeft[0], topRight[0], self.resolutionX)
         yRange = np.linspace(topLeft[1], bottomLeft[1], self.resolutionY)
 
         # array to store hit information
-        values = np.empty((xRange.size, yRange.size), dtype=object)
+        values = np.empty((xRange.size, yRange.size, 3), dtype=object)
         labels = np.empty((xRange.size, yRange.size), dtype=object)
 
         # indices for array mapping
         indexX = 0
         indexY = 0
-
         # iterate over all X/Y coordinates
         for x in xRange:
             for y in yRange:
@@ -164,20 +211,24 @@ class Plant_Processor:
                 pixelVector = Vector((x, y, topLeft[2]))
                 # rotate that vector according to camera rotation
                 pixelVector.rotate(cam.matrix_world.to_quaternion())
-                origin = cam.location
+                #pixelVector.rotate(container.matrix_world.to_quaternion())
+                #destination = matrixWorldInverted @ (pixelVector + cam.matrix_world.translation)
+                #direction = (destination - origin).normalized()
+                origin = cam.matrix_world.to_translation()
 
                 # perform the actual ray casting
                 hit, location, norm, face, hit_object, matrix = bpy.context.scene.ray_cast(bpy.context.view_layer,origin,pixelVector)
 
                 if hit:
-                    values[indexX,indexY] = location
-                    labels[indexX,indexY] = hit_object.name.split('.')[0]
+                    values[indexX,indexY] = np.array(location)
+                    labels[indexX,indexY] = label_dict[hit_object.name.split('.')[0]]
 
                 # update indices
                 indexY += 1
 
             indexX += 1
             indexY = 0
+
         return values, labels
 
     def show_point_cloud(self, values):
@@ -185,10 +236,12 @@ class Plant_Processor:
         mesh = bpy.data.meshes.new(name='created mesh')
         bm = bmesh.new()
 
+        vls = values.reshape(values.shape[0]*values.shape[1],3)
+
         # iterate over all possible hits
-        for index, location in np.ndenumerate(values):
+        for index, location in enumerate(vls):
             # no hit at this position
-            if location is not None:
+            if location[0] is not None:
                 # add new vertex
                 bm.verts.new((location[0], location[1], location[2]))
 
@@ -212,7 +265,6 @@ class Plant_Processor:
         bpy.ops.object.select_all(action='DESELECT')
         obj.select_set(True)
         bpy.context.view_layer.objects.active = obj
-
         # reset view mode
         #bpy.context.area.type = mode
 
@@ -222,7 +274,7 @@ class Plant_Processor:
         for split_object in bpy.data.objects: # Check for given object names
             bpy.ops.object.select_all(action='DESELECT') # deselect all objects
 
-            if not 'Camera' in split_object.name and not 'Light' in split_object.name:
+            if not 'Camera' in split_object.name and not 'Light' in split_object.name and not 'Empty' in split_object.name:
                 split_object.select_set(True) #only select one at a time
                 dae_save_string = '%s_%s.dae' %(i,split_object.name.split('.')[0])
                 obj_save_string = '%s_%s.obj' %(i,split_object.name.split('.')[0])
@@ -256,3 +308,4 @@ if __name__ == "__main__":
     pproc = Plant_Processor()
     pproc.init_cleanup() #First clean the scene
     pproc.open_and_split() #Load plant objects in a loop and perform operations on them
+    print("Data processing completed")
